refactor: extract SQLiteDatabaseWriter to src/refactors/sqlite_database_writer.py (SC-003)
Extracts the 316 LoC SQLiteDatabaseWriter class from MistHelper.py into
src/refactors/sqlite_database_writer.py. Uses late-import dependency
injection (importlib + SimpleNamespace) to avoid circular imports with
DatabaseSchemaUtils and DataProcessingUtils, and resolves DATABASE_PATH
at call-time to preserve monkeypatch support in tests.

Decomposes two functions per FR-006 while moving:
- _create_table_and_indexes -> parent + _create_schema_table + _create_schema_indexes
- _insert_single_row       -> parent + _log_sample_insert + _log_row_failure

Extracted file compliance: 100.0/A+ (32 functions, avg CC 2.4, max CC 5).
Repo baseline preserved at 99.5/A+.

Word-boundary grep confirms only two SQLiteDatabaseWriter references
outside the extracted file: the new import at MistHelper.py:149 and the
single call site at MistHelper.py:7096.

Ruff auto-removed unused sqlite3 import from MistHelper.py; the module
is imported by the extracted class where it is actually used.

Implements T027-T037.
