+++ .specify/feature.json
@@ -1,1 +1,1 @@
-{"feature_directory": "specs/1009-compliance-backlog-remediation"}
+{"feature_directory": "specs/1010-misthelper-refactor-extraction"}
